Split search response into smaller files
