+++ src/search/response/inner_hits_result.rs
@@ -1,0 +1,8 @@
+use super::HitsMetadata;
+
+/// Represents inner hits
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct InnerHitsResult {
+    /// The actual inner hits
+    pub hits: HitsMetadata,
+}
