+++ src/search/response/shard_failure.rs
@@ -1,0 +1,2 @@
+/// What went wrong
+pub type ShardFailure = serde_json::Value;
